Folder structure 2nd iteration
PCOR-xxx

Removed inc and src folders
