+++ application/CMakeLists.txt
@@ -9,6 +9,6 @@ add_executable (
 
 target_link_libraries (
     console_app
-    core
+    invoker
     utility
 )
