+++ invoker/domain/controller/authcontroller.hpp
@@ -18,43 +18,29 @@
 *           Ben Ziv <pointonsoftware@gmail.com>                                                   *
 *                                                                                                 *
 **************************************************************************************************/
-#include <entity/person.hpp>
+#ifndef DOMAIN_PUBLIC_CONTROLLER_AUTHCONTROLLER_HPP_
+#define DOMAIN_PUBLIC_CONTROLLER_AUTHCONTROLLER_HPP_
+
+#include <string>
+#include <memory>
+#include <domain/defines.hpp>
+#include <domain/interface/view/authviewif.hpp>
+#include <domain/interface/dataprovider/authdataif.hpp>
 
 namespace domain {
-namespace entities {
+namespace authentication {
 
-Person::Person(const std::string& firstname,
-               const std::string& middlename,
-               const std::string& lastname,
-               const std::string& birthdate,
-               const std::string& gender)
-: m_firstname(firstname), m_middlename(middlename), m_lastname(lastname),
-  m_birthdate(birthdate), m_gender(gender) {
-    // Empty for now
-}
+class AuthController {
+ public:
+    explicit AuthController(std::unique_ptr<AuthViewIface>&& view,
+                            std::unique_ptr<AuthDataProviderIface>&& dataprovider);
+    bool login(const std::string& pin);
+ private:
+    status::General authenticate(const std::string& pin);
+    std::unique_ptr<AuthViewIface> mView;
+    std::unique_ptr<AuthDataProviderIface> mDataProvider;
+};
 
-void Person::addPhoneNumber(const std::string& phonenumber) {
-    m_contact_details.phone_number.emplace_back(phonenumber);
-}
-
-void Person::addPersonalId(const std::string& type, const std::string& number) {
-    m_personal_ids.emplace_back(PersonalId{type, number});
-}
-
-void Person::setPersonalIds(const std::vector<PersonalId>& personalids) {
-    if (!m_personal_ids.empty()) {
-        // warning, we're overwriting the container
-    }
-    m_personal_ids = personalids;
-}
-
-void Person::setEmail(const std::string& email) {
-    m_contact_details.email = email;
-}
-
-void Person::setAddress(const Address& address) {
-    m_address = address;
-}
-
-}  // namespace entities
+}  // namespace authentication
 }  // namespace domain
+#endif  // DOMAIN_PUBLIC_CONTROLLER_AUTHCONTROLLER_HPP_
