+++ utility/logclient/CMakeLists.txt
@@ -1,0 +1,7 @@
+project (logclient)
+
+# Logger client
+add_executable (
+    log_client_app
+    logclient.cpp
+)
