+++ utility/logclient/logclient.cpp
